Fixed obvious provenance issues

+++ include/alloc_buffer.h
@@ -163,7 +163,7 @@ alloc_buffer_mark_failed (struct alloc_buffer *buf)
 static __always_inline __attribute__ ((nonnull (1))) size_t
 alloc_buffer_size (const struct alloc_buffer *buf)
 {
-  return buf->__alloc_buffer_end - buf->__alloc_buffer_current;
+  return (size_t)buf->__alloc_buffer_end - (size_t)buf->__alloc_buffer_current;
 }
 
 /* Return true if the buffer has been marked as failed.  */
@@ -264,9 +264,9 @@ __alloc_buffer_alloc (struct alloc_buffer *buf, size_t size, size_t align)
   if (size == 1 && align == 1)
     return alloc_buffer_alloc_bytes (buf, size);
 
-  size_t current = buf->__alloc_buffer_current;
-  size_t aligned = roundup (current, align);
-  size_t new_current = aligned + size;
+  uintptr_t current = buf->__alloc_buffer_current;
+  uintptr_t aligned = roundup (current, align);
+  uintptr_t new_current = aligned + size;
   if (aligned >= current        /* No overflow in align step.  */
       && new_current >= size    /* No overflow in size computation.  */
       && new_current <= buf->__alloc_buffer_end) /* Room in buffer.  */
@@ -298,8 +298,8 @@ __alloc_buffer_next (struct alloc_buffer *buf, size_t align)
   if (align == 1)
     return (const void *) buf->__alloc_buffer_current;
 
-  size_t current = buf->__alloc_buffer_current;
-  size_t aligned = roundup (current, align);
+  uintptr_t current = buf->__alloc_buffer_current;
+  uintptr_t aligned = roundup (current, align);
   if (aligned >= current        /* No overflow in align step.  */
       && aligned <= buf->__alloc_buffer_end) /* Room in buffer.  */
     {
